+++ src/main/java/me/shepherd23333/projecte/emc/mappers/customConversions/json/FixedValues.java
@@ -3,6 +3,7 @@ package me.shepherd23333.projecte.emc.mappers.customConversions.json;
 import com.google.gson.annotations.SerializedName;
 import me.shepherd23333.projecte.emc.json.NormalizedSimpleStack;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -10,8 +11,8 @@ import java.util.Map;
 
 public class FixedValues {
     @SerializedName("before")
-    public Map<NormalizedSimpleStack, Long> setValueBefore = new HashMap<>();
+    public Map<NormalizedSimpleStack, BigInteger> setValueBefore = new HashMap<>();
     @SerializedName("after")
-    public Map<NormalizedSimpleStack, Long> setValueAfter = new HashMap<>();
+    public Map<NormalizedSimpleStack, BigInteger> setValueAfter = new HashMap<>();
     public List<CustomConversion> conversion = new ArrayList<>();
 }
